Replace and rename the acknowledgements objects
