+++ lib/streamy/consumer.rb
@@ -3,7 +3,7 @@ require "hutch"
 module Streamy
   module Consumer
     def self.included(base)
-      base.include Hutch::Consumer
+      base.include ::Hutch::Consumer
       base.consume "#{Streamy::DEFAULT_TOPIC_PREFIX}.#"
 
       configure_hutch
@@ -14,9 +14,12 @@ module Streamy
     end
 
     def self.configure_hutch
-      Hutch::Config.set(
+      ::Hutch::Config.set(
         :error_acknowledgements,
-        [MessageBuses::RabbitMessageBus::Resquer.new, MessageBuses::RabbitMessageBus::Aborter.new]
+        [
+          Hutch::Acknowledgements::ResqueOnAllFailures.new,
+          Hutch::Acknowledgements::AbortOnAllFailures.new
+        ]
       )
     end
   end
