+++ lib/streamy/hutch/acknowledgements/abort_on_all_failures.rb
@@ -1,0 +1,15 @@
+require "hutch"
+
+module Streamy
+  module Hutch
+    module Acknowledgements
+      class AbortOnAllFailures < ::Hutch::Acknowledgements::Base
+        def handle(*)
+          puts "[x] abort consumer"
+
+          exit
+        end
+      end
+    end
+  end
+end
