+++ lib/streamy/hutch/acknowledgements/reque_on_all_failures.rb
@@ -1,0 +1,16 @@
+require "hutch"
+
+module Streamy
+  module Hutch
+    module Acknowledgements
+      class ResqueOnAllFailures < ::Hutch::Acknowledgements::Base
+        def handle(delivery_info, properties, broker, ex)
+          broker.requeue(delivery_info.delivery_tag)
+          puts "[*] requeued message(#{properties.message_id || '-'})"
+
+          false
+        end
+      end
+    end
+  end
+end
