--- lib/streamy/message_buses/rabbit_message_bus/aborter.rb
@@ -1,13 +1,0 @@
-require "hutch"
-
-module Streamy
-  module MessageBuses
-    class RabbitMessageBus::Aborter < Hutch::Acknowledgements::Base
-      def handle(*)
-        puts "[x] abort consumer"
-
-        exit
-      end
-    end
-  end
-end
