--- lib/streamy/message_buses/rabbit_message_bus/resquer.rb
@@ -1,14 +1,0 @@
-require "hutch"
-
-module Streamy
-  module MessageBuses
-    class RabbitMessageBus::Resquer < Hutch::Acknowledgements::Base
-      def handle(delivery_info, properties, broker, ex)
-        broker.requeue(delivery_info.delivery_tag)
-        puts "[*] requeued message(#{properties.message_id || '-'})"
-
-        false
-      end
-    end
-  end
-end
